feat: component admin page + remove auto intake
